requirements.txt
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
cachetools==5.5.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
flake8==7.1.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
groq==0.10.0
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.6
identify==2.6.0
idna==3.8
importlib_metadata==8.4.0
jiter==0.5.0
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
mccabe==0.7.0
nodeenv==1.9.1
openai==1.42.0
packaging==24.1
pdoc3==0.11.1
platformdirs==4.2.2
pre-commit==3.8.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
pyparsing==3.1.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
zipp==3.20.1
